+++ src/main/kotlin/org/jetbrains/intellij/dependency/IdeaExtraDependency.kt
@@ -1,6 +1,5 @@
 package org.jetbrains.intellij.dependency
 
-
 import org.jetbrains.intellij.collectJars
 import java.io.File