+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - Use Gradle `ArchiveOperations` in `extractArchive` utils method [#681](../../issues/681)
 - Set minimal supported Gradle version to 6.6
+- Use JDOM for altering `updates.xml` in `PrepareSandboxTask` to keep existing content
 
 ## 1.0.0