+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Use Gradle `ArchiveOperations` in `extractArchive` utils method [#681](../../issues/681)
 - Set minimal supported Gradle version to 6.6
 - Use JDOM for altering `updates.xml` in `PrepareSandboxTask` to keep existing content
+- Fixed incorrect output path of JarSearchableOptionsTask causing also duplicate entry exception [#678](../../issues/678)
 
 ## 1.0.0