+++ CHANGES.md
@@ -12,6 +12,7 @@
 - `BuildSearchableOptionsTask` fails on MacOS when resolving `javaHome` #[#696](../../issues/696)
 - `PrepareSandboxTask` doesn't depend on `JavaPlugin` dependencies #[#451](../../issues/451)
 - Remove `IntelliJPluginExtension.pluginsRepositories(block: Closure<Any>)` due to `ConfigureUtil` deprecation and a lack of typed parameters
+- Remove usage of deprecated methods and classes introduced in Gradle 7.1 #[#700](../../issues/700)
 
 ## 1.0.0