+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 1.0.1, not published
+## 1.1.0
 
 - Use Gradle `ArchiveOperations` in `extractArchive` utils method [#681](../../issues/681)
 - Set minimal supported Gradle version to 6.6